requirements.txt
boto3
aws-cdk-lib==2.55.1
constructs>=10.0.0,<11.0.0
